tools/competitor-scraper/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "competitor-scraper"
version = "0.1.0"
description = "EBQ competitor crawler — admin-triggered Scrapy worker that walks a competitor domain, extracts keywords + topics, and writes back to the EBQ MySQL database."
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [{ name = "EBQ" }]

dependencies = [
    "scrapy>=2.11,<3",
    "lxml>=5",
    "readability-lxml>=0.8.1",
    "yake>=0.4.8",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "sqlalchemy>=2.0",
    "pymysql>=1.1",
    "python-dotenv>=1",
    "pydantic>=2.5",
    "pydantic-settings>=2",
    "typer>=0.12",
    "tldextract>=5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
]

[project.scripts]
competitor-scraper = "competitor_scraper.cli:app"

[tool.setuptools.packages.find]
include = ["competitor_scraper*"]
exclude = ["tests*"]
